feat(package.json): upgrade flowise-embed to ^1.3.7 refactor(_app.tsx): replace BubbleChat with Script tag to load flowise-embed
The `flowise-embed` package has been upgraded to version `^1.3.7` to incorporate the latest features and bug fixes. In `_app.tsx`, the `BubbleChat` component from `flowise-embed-react` has been removed in favor of a `Script` tag that loads the new `flowise-embed` directly. This change simplifies the integration by removing an unnecessary package dependency and allows for more flexible initialization of the chatbot with additional configuration options directly within the script.